requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
async-generator==1.10
attrs==22.1.0
audioread==3.0.0
autopep8==2.0.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.6
cp-tool==2.1.2
cryptography==38.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
djangorestframework==3.14.0
duplicity==0.8.21
et-xmlfile==1.1.0
excel2json-3==0.1.6
exceptiongroup==1.0.0rc9
fasteners==0.14.1
Flask==2.0.1
fonttools==4.38.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.3
grequests==0.6.0
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.2.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librosa==0.9.2
llvmlite==0.39.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pooch==1.6.0
protobuf==3.12.4
ptyprocess==0.7.0
PyAudio==0.2.12
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
PyGithub==1.55
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.27.1
resampy==0.4.2
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.1
selenium==4.2.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.4
systemd-python==234
threadpoolctl==3.1.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webdriver-manager==3.7.1
Werkzeug==2.0.2
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
